Animations Wired and Coded Up
The code in AIPlayer has been changed so now the animations should be lined up with where the code tells the AI to place a card. The movement or timing of the AI card hasn't been changed yet to work with the current timing of animations, but other than that the animations should be wired and coded correctly now.