Card Moves with Animations
I have tried to have the card move in sync with the animations, and it's currently working, although a bit more polish may be needed. I have edited the code of AIPlayer to move the card based on Animation Events, which are points in an animation that you can run a script function.