Card Guide Fix
The card guides would usually flicker for about a frame when the next card spawned. I've implemented a fix so they shouldn't do that anymore. I also added a second card screen in the MultiMode scene that shows what Player 1's next card is as well, should we decide to still use them in the MultiMode scene.